Various bashrc improvements
The main thing that changed here is that I figured out how to get fg to
reset the terminal to the preferred size if it's reviving vim or
vimdiff.  Along the way, I decided that I want to use explicit "-n" in
my tests, and that I wanted to abstract my terminal escape sequences
into two functions `term-set-cols` and `term-set-caption`

Also, when I did that last one, I realized that vim's default syntax
file for type 'sh' doesn't consider '-' to be part of a keyword, so I
had to copy it into my `.vim` and update it.