Newline issue in global_config
Had to make an actual change so that I could commit the file after
running dos2unix on it; that was the previous commit. This one undoes
that change, so that it's the same as two commits ago except it has lf
only instead of crlf.

It seems I completely bungled the global_config stuff when I copied it
over from a different directory.