Updated grep alias
I am no longer overwriting the `grep` command; instead I'm creating the
`pgrep` alias (in the style of `egrep` or `fgrep`) that uses the most
advanced regexp style (perl or falls back to extended) available on the
system.

Also in this commit, I have tweaked the `lld*` family of aliases to
preserve ls's coloring through the grep piping.