SSH config change
SSH uses the first value it encounters in its config files, so the most
general settings need to go at the end. This means that the * host needs
to be at the end of the global_config file, and the local_config file
needs to be cat'd into config before global_config.